Tutorial, attack sounds, and winscreen
Added three main things, a little tutorial that explains the controls and what each thing is/does in the level, added sound effects for each attack, and added a Win screen that gets called after the last wave. I also did a backup of the scenes I edited incase they get overwritten in a merge, they're in their own folder.